refactor: Community 게시글 삭제 시 이미지 처리 수정 (#70)

+++ api-module/src/main/java/org/devridge/api/domain/community/service/CommunityService.java
@@ -76,8 +76,10 @@ public class CommunityService {
         updateByHashtagIds(communityHashtags);
 
         communityRepository.deleteById(communityId);
-        List<String> images = Arrays.asList(community.getImages().split(", "));
-        s3Service.deleteAllImage(images);
+        if (community.getImages() != null) {
+            List<String> images = Arrays.asList(community.getImages().split(", "));
+            s3Service.deleteAllImage(images);
+        }
     }
 
     public List<CommunityListResponse> getAllCommunity() {
